% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SummaryZib}
\alias{SummaryZib}
\title{Extract and summarize posterior distribution}
\usage{
SummaryZib(obj, iter)
}
\arguments{
\item{obj}{model object from \code{zoib} where \code{one.inflation = FALSE}
and \code{joint = FALSE}.}

\item{iter}{integer. The number of iterations post-burn in.}
}
\value{
data frame
}
\description{
Extract and summarize posterior distribution
}

