% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zibPlot}
\alias{zibPlot}
\title{Function to plot parameter posteriors from Zero-inflated Beta Regressions}
\usage{
zibPlot(obj, max, variable_names = NULL, xlab = "\\nCoef. Estimates",
  title = "")
}
\arguments{
\item{obj}{model object from \code{zoib} where \code{one.inflation = FALSE}
and \code{joint = FALSE}.}

\item{max}{maximum number of simulations post burn-in}

\item{variable_names}{character vector of variable names. Must be in the
same order and the same length as variables enterd into the \code{zoib}
formula.}

\item{xlab}{character string x-axis label.}

\item{title}{character string title for the plot}
}
\value{
a ggplot2 plot
}
\description{
Function to plot parameter posteriors from Zero-inflated Beta Regressions
}

