% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GelmanDiag}
\alias{GelmanDiag}
\title{Extract Gelman-Rubin diagnostics as data frame}
\usage{
GelmanDiag(obj, iter)
}
\arguments{
\item{obj}{model object from \code{zoib} where \code{one.inflation = FALSE}
and \code{joint = FALSE}.}

\item{iter}{integer. The number of iterations post-burn in.}
}
\value{
data frame
}
\description{
Extract Gelman-Rubin diagnostics as data frame
}

