% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SummaryZib}
\alias{SummaryZib}
\title{Extract and summarise posterior distribution}
\usage{
SummaryZib(obj, iter)
}
\arguments{
\item{obj}{model object from \code{zoib} where \code{one.inflation = FALSE}
and \code{joint = FALSE}.}

\item{iter}{integer. The number of iterations.}
}
\value{
data frame
}
\description{
Extract and summarise posterior distribution
}
\examples{
\dontrun{
Load packages
library(zoib)
library(zibHelpers)

# Run example of clustered zero-inflated beta regression from
# Liu and Kong (under review, 12-14)

# Load data
data("AlcoholUse", package = "zoib")
AlcoholUse$Grade <- as.factor(AlcoholUse$Grade)

nIter = 50 # Number of iterations, including burn-in

# Estimate
Out <- zoib(Percentage ~ Grade + Days + Gender|1|Grade + Days + Gender|1,
            data = AlcoholUse, random = 1, EUID = AlcoholUse$County,
            zero.inflation = TRUE, one.inflation = FALSE, joint = FALSE,
            n.iter = nIter)

# Summarise the posterior
SummaryZib(Out, nIter)
}
}

